% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_doy.internal.r
\name{axis_doy.internal}
\alias{axis_doy.internal}
\title{Create custom axis starting on hydrologic year start month}
\usage{
axis_doy.internal(hyrstart = 10)
}
\arguments{
\item{hyrstart}{numeric indicating month for start of the hydrologic year 
(water year).}
}
\description{
Create custom axis starting on hydrologic year start month
}
\author{
Paul Whitfield
}
