% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.station.internal.r
\name{get.station.internal}
\alias{get.station.internal}
\title{Get station information for hydrometric stations}
\usage{
get.station.internal(Agency, StationID)
}
\arguments{
\item{Agency}{Character string of Agency.}

\item{StationID}{Character string of station ID.}
}
\value{
Returns a data.frame of station information
}
\description{
Get station information for ROBIN, USGS, or WSC hydrometric stations.
}
\author{
Jennifer Dierauer
}
