% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.titles.internal.R
\name{get.titles.internal}
\alias{get.titles.internal}
\title{Returns plot titles and labels based on plot type and language preference}
\usage{
get.titles.internal(type, flow.units, language = "English", Qmax)
}
\arguments{
\item{type}{character indicating the type of summary plot}

\item{flow.units}{Character string indicating the units for streamflow values, one of either
'ft3/s' or 'm3/s', taken from the flow time series data.frame created with \code{read.flows}}

\item{language}{"English" or "French"}

\item{Qmax}{the flow quantile used to define peaks of threshold, e.g. 0.95}
}
\description{
Returns plot titles and labels based on plot type and language preference
}
\author{
Jennifer Dierauer
}
