% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyear.internal.r
\name{hyear.internal}
\alias{hyear.internal}
\title{Add hydrologic Year, month, and doy columns to a daily time series}
\usage{
hyear.internal(TS, hyrstart = 10)
}
\arguments{
\item{TS}{Output from \code{\link{create.ts}} function.}

\item{hyrstart}{define start month of hydrologic year. Defaults to 10 (October).}
}
\value{
Returns a data.frame with hyear, hmonth, and hdoy columns 
  appended to the original input data.frame.
}
\description{
Add hydrologic Year, month, and doy columns to a daily time series
}
\author{
Jennifer Dierauer
}

