% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.frames.internal.R
\name{screen.frames.internal}
\alias{screen.frames.internal}
\title{Internal wrapper for creating trend and change-point plots}
\usage{
screen.frames.internal(
  input,
  mparam,
  mylab,
  DataType,
  maf,
  mmar,
  text,
  xaxis,
  Year1,
  YearEnd,
  hyrstart
)
}
\arguments{
\item{input}{metric time series}

\item{mparam}{trend and change point info}

\item{mylab}{y axis label}

\item{DataType}{numeric indicating data type}

\item{maf}{mean annual flow series}

\item{mmar}{plot margins}

\item{text}{title passed from screen.frames (text or NULL)}

\item{xaxis}{boolean indicating whether to plot the x axis}

\item{Year1}{start year of original time series}

\item{YearEnd}{end year of original time series}

\item{hyrstart}{numeric indicating month for start of the hydrologic year}
}
\description{
Internal wrapper for creating trend and change-point plots
}
\author{
Jennifer Dierauer
}
