% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.summary.internal.R
\name{screen.summary.internal}
\alias{screen.summary.internal}
\title{Internal wrapper for creating trend and change-point summary plots}
\usage{
screen.summary.internal(input, mparam, type, ylabs, i, DataType, maf, Year1,
  YearEnd, hyrstart)
}
\arguments{
\item{input}{metric time series}

\item{mparam}{trend and change point info}

\item{type}{character indicating type of summary plot}

\item{ylabs}{y axis labels}

\item{i}{plot position}

\item{DataType}{numeric indicating data type}

\item{maf}{mean annual flow series}

\item{Year1}{start year of original time series}

\item{YearEnd}{end year of original time series}

\item{hyrstart}{numeric indicating month for start of the hydrologic year}
}
\description{
Internal wrapper for creating trend and change-point summary plots
}
\author{
Jennifer Dierauer
}
