% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.station.internal.r
\name{get.station.internal}
\alias{get.station.internal}
\title{Get station information for USGS or WSC hydrometric stations}
\usage{
get.station.internal(stnID)
}
\arguments{
\item{stnID}{Character string of station ID.}
}
\value{
Returns a list of station information
}
\description{
Get station information for USGS or WSC hydrometric stations
}
\author{
Jennifer Dierauer
}
