% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.titles.internal.R
\name{get.titles.internal}
\alias{get.titles.internal}
\title{Returns plot titles and labels based on plot type and language preference}
\usage{
get.titles.internal(type, language = "English", Qmax)
}
\arguments{
\item{type}{character indicating the type of summary plot}

\item{language}{"English" or "French"}

\item{Qmax}{the flow quantile used to define peaks of threshold, e.g. 0.95}
}
\description{
Returns plot titles and labels based on plot type and language preference
}
\author{
Jennifer Dierauer
}

