% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.flows.R
\name{read.flows}
\alias{read.flows}
\title{Read .csv or .Rdata file of streamflows}
\usage{
read.flows(filename, convert = TRUE)
}
\arguments{
\item{filename}{name of .csv file to be read.}

\item{convert}{Boolean indicating whether or not to convert USGS flow values
from cubic feet per second to cubic meters per second.}
}
\description{
Reads .csv or .Rdata files of daily streamflow time series.  Recognizes 
several formats, including those used by Water Survey Canada and the United 
States Geological Survey. Uses read.csv() or load() functions from base package, 
returns data frame with ID, Date, and Flow, and, if available,
associated quality codes and source agency. Replaces negative values that are 
sometimes used to denote missing data with NAs.
}
\author{
Jennifer Dierauer
}

