% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YMD.internal.r
\name{YMD.internal}
\alias{YMD.internal}
\title{Add calendar year, month, and day of year columns}
\usage{
YMD.internal(TS)
}
\arguments{
\item{TS}{Output from \code{\link{create.ts}} function.}
}
\value{
Returns a data.frame with year, month, and doy columns appended.
}
\description{
Add calendar year, month, and day of year columns
}
\author{
Jennifer Dierauer
}
